# kawamonn-storage/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Kawamonn Storage & SSH Platform API
  version: 1.0.0
  description: REST API for managing online storage and SSH environments.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Backend Service
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email:
                  type: string
                password:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  status:
                    type: string
  /auth/verify:
    post:
      summary: Verify email code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_name, password]
              properties:
                account_name:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
  /files:
    post:
      summary: Upload a file (multipart)
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  size:
                    type: integer
                  mime_type:
                    type: string
    get:
      summary: List files
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  total_count:
                    type: integer
  /files/{id}:
    get:
      summary: Get file stream or preview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
  /files/{id}/share:
    post:
      summary: Issue share link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expires_at:
                  type: string
                  format: date-time
                password:
                  type: string
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  share_token:
                    type: string
  /admin/users:
    get:
      summary: List all users (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: Add exceptional user (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, account_name, quota]
              properties:
                email:
                  type: string
                account_name:
                  type: string
                quota:
                  type: integer
      responses:
        '201':
          description: User added
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
  /admin/users/{id}:
    delete:
      summary: Delete user (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /ssh/provision:
    post:
      summary: Provision SSH Container
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Container provisioned
          content:
            application/json:
              schema:
                type: object
                properties:
                  container_id:
                    type: string
                  status:
                    type: string
  /ssh/status/{username}:
    get:
      summary: Get SSH container status
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  container_status:
                    type: string
                  resource_usage:
                    type: object
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT